fetch/metadata/generated/form-post-sec-fetch-mode.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-Mode</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(path) {
    const win = window.open();
    const inputs = [
        ['redirect_status', '307'],
        ['body', `<script>opener.postMessage('done', '*');</${''}script>`],
        ...new URL(path, location.origin).searchParams
      ]
      .map(([name, value]) =>
        `<input type="hidden" name="${name}" value="${value}">`
      )
      .join('\n');

    return new Promise((resolve) => {
        addEventListener('message', function(event) {
          if (event.source === win) {
            resolve();
          }
        });
        win.document.open();
        win.document.write(`
          <!DOCTYPE html>
          <form method="POST" action="${path}">${inputs}</form>
          <script>
            const form = document.getElementsByTagName('form')[0];
            form.submit();
          </${''}script>
        `);
        win.document.close();
      })
      .then(() => win.close());
  }

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';
    return induceRequest('' + path)
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'navigate');
        });
  }, 'sec-fetch-mode - ');
  </script>
  </body>
</html>
